# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/warehouse/status": {
            "post": {
                "description": "Update status warehouse",
                "produces": [
                    "application/json"
                ],
                "summary": "Update status warehouse",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/request.WarehouseStatus"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Product"
                        }
                    }
                }
            }
        },
        "/warehouse/transfer": {
            "post": {
                "description": "Transfer Stock of product between warehouse",
                "produces": [
                    "application/json"
                ],
                "summary": "Transfer Stock of product between warehouse",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/request.WarehouseTransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Product"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Product": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "stock_quantity": {
                    "type": "integer"
                },
                "theme": {
                    "type": "string"
                }
            }
        },
        "request.WarehouseStatus": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                }
            }
        },
        "request.WarehouseTransferRequest": {
            "type": "object",
            "properties": {
                "from_warehouse_id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "to_warehouse_id": {
                    "type": "integer"
                }
            }
        }
    }
}